ml/requirements.txt
grpcio
grpcio-tools
miniopy-async
ffmpeg-python
scikit-learn
numpy
pandas
scipy
python-dotenv
opencv-python
acapture
aiofiles
pymongo
pillow
catboost